requirements.txt
anyio==3.6.1
asgiref==3.5.0
click==8.1.3
colorama==0.4.5
dj-database-url==0.5.0
Django==4.0.4
django-froala-editor==4.0.12
django-heroku==0.3.1
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
mypy==0.961
mypy-extensions==0.4.3
mypy-lang==0.5.0
mysql-connector-python==8.0.29
Pillow==9.2.0
psycopg2==2.9.3
pydantic==1.9.1
python-decouple==3.6
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
uvicorn==0.18.2
whitenoise==6.2.0
